Version 1
This code is a YouTube video downloader app created using Python and PyQt5. The app can download either a single YouTube video or a playlist, and offers the user the option to select the video quality and whether to download only the audio. The user can input the video or playlist URL, select the download options, and initiate the download with a click of a button.